refactor: move CashuToken and Proof to commons
Move CashuToken data class and Proof serializable class from
amethyst.service.cashu to commons.data.cashu for KMP shared access.
Add kotlinx-serialization plugin to commons module.
Update all consumers to import from commons directly.
