+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/data/cashu/CashuToken.kt
@@ -18,7 +18,24 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.service.cashu
+package com.vitorpamplona.amethyst.commons.data.cashu
 
-typealias CashuToken = com.vitorpamplona.amethyst.commons.data.cashu.CashuToken
-typealias Proof = com.vitorpamplona.amethyst.commons.data.cashu.Proof
+import androidx.compose.runtime.Immutable
+import kotlinx.serialization.Serializable
+
+@Immutable
+data class CashuToken(
+    val token: String,
+    val mint: String,
+    val totalAmount: Long,
+    val proofs: List<Proof>,
+)
+
+@Serializable
+@Immutable
+class Proof(
+    val amount: Int,
+    val id: String,
+    val secret: String,
+    val C: String,
+)
